Updated to avoid visible plain text passwords in console
I am newer to Azure; however, I know that on typical LAMP systems we are religious about not putting our passwords in plain text into the console, as it can be easily viewed. I figure this practice would be the same with Azure, so I opted to modify the tutorial a little. I tested it first to confirm that it actually prompts for the password.